slave: fix user switching for consolekit users
ConsoleKit has code to try to prevent switching to the same session
that the slave is running on.  That code was important when we created
a new throw-away session and slave as part of the user switching process.

These days we connect to the already running slave instead of using a
new one, and that vestigal code gets in the way, breaking user
switching.

This commit removes that harmful code.

https://bugzilla.gnome.org/show_bug.cgi?id=684594