launch-environment: start dbus-daemon inside the user session
We currently start dbus-daemon separately from the launch
environment PAM session.  This means all things activated
by that bus get started in the wrong logind session and are
missing important environment variables (such as XDG_RUNTIME_DIR).

This commit deletes a bunch of code for managing the dbus session
separately, and instead just makes it part of the session command
(using dbus-launch as a wrapper around the session).

Because we no longer have the specific PID of the bus daemon,
we now stop the launch environment by killing the whole process
group in one go.

https://bugzilla.gnome.org/show_bug.cgi?id=684474