daemon: user marker file for tracking autologin instead of static variable.
We currently decide whether or not to skip autologin by a first_login
state variable in the static display object.

These days we can have multiple static display objects, so storing the
state variable doesn't make much sense.  We could make the variable
static, but instead this commit switches to using a marker file in
/var/run/gdm.

https://bugzilla.gnome.org/show_bug.cgi?id=682467