slave: don't fail if user session already active when activating from unlock
After GDM successfully reauthenticates a user it runs

gdm_slave_switch_to_user_session

to jump to the user's VT and unlock it.  Of course, if the user
is reauthenticating from an unlock screen instead of a user
switched login screen, then they'll already be on the right VT.

gdm_slave_switch_to_user_session fails if the user is already
switched to the session specified.  This failing behavior is used
at login time to decide start a user's session (versus switching
to a user's already running session).

This commit changes gdm_slave_switch_to_user_session to take
an additional parameter to decide whether or not switching to
an already active session should fail, or merely skip to the
unlocking step.

https://bugzilla.gnome.org/show_bug.cgi?id=685988