slave: cancel start session request if slave gets stopped
If the slave gets stopped right before we start the user's session,
then we'll have a queued start session in flight, and variable set
to the session name to be started duplicated. This gunks up the
stopping logic and leads to infinite recursion.

This commit cancels the start session operation to correct the bug.

https://bugzilla.gnome.org/show_bug.cgi?id=689367