session: Don't accept empty string as saved language
accountsservice currently returns the empty string for Language (and
other string keys) when they're unset, instead of NULL.  GDM should be
defensive against this.

This way we get a proper fallback to the system's default locale.
Fixes setting of $LANG in user session in gnome-ostree buildmaster.

See also:
http://lists.freedesktop.org/archives/systemd-devel/2013-January/007945.html
for a discussion of the current state of $LANG, pam, systemd, gdm, and
accountsservice.

https://bugzilla.gnome.org/show_bug.cgi?id=699261