slave: don't assume autologin if there are no connections
Because of a bug in gdm-session.c (which was dealt with in
the previous commit) we ended up with a false impression of currently
connected clients after a session reset happens.

That bug meant the slave assumed autologin was enabled which
stopped it from waiting for the greeter to finish the fade out before
proceeding.

Assuming "no connection" == "autologin" is a bit fragile, so this
commit makes sure to check for gdm-autologin specifically.

https://bugzilla.gnome.org/show_bug.cgi?id=703386