session: harden against server dbus outliving session object
Right now we do:

g_signal_connect (dbus_server, "...", callback, self);

in a few places.  These signal connections should go away when the dbus
server is unreffed in dispose, but to be on the safe side we should
use g_signal_connect_object anyway.  (See for instance the leak
fix in the previous commit).