worker: give hint about password expiration
When you run "passwd -f (username)" to expire a users password, and then try to
login via GDM the following happens:

- user enters their username
- user enters their password
- The dialog prompt says "New password"
- The user enters a new password
- The dialog prompt says "Enter password again"
- The user enters the same password, changes their password and their session
  starts.

Some of our users were confused why the dialog was asking for a new password.

This commit adds code to show the user a message explaining that the user
needs to enter their new password.

https://bugzilla.gnome.org/show_bug.cgi?id=629713