daemon: support launching sessions on separate VT
If an xsession file sets X-GDM-NeedsVT to true, then GDM will
now automatically allocate a new VT for that session and jump to
it before starting the session.  Once the session completes, the
worker will automatically jump back to the VT it started on.

This will be useful for getting mutter-launch to be able to launch
a gnome-shell wayland sesssion.  Longer term, I think mutter-launch
will go away and its functionality will get moved into logind.