slave: don't kill slave when client disconnects
Right now we stop the slave if a greeter running on a remote
X server exits. This is wrong, since the greeter will exit as
part of the login screen process.  We already have mechanisms in
place to reap the slave at the appropriate time, so stopping the
slave on disconnection is totally unneeded and actively harmful.

This commit drops that code.

https://bugzilla.gnome.org/show_bug.cgi?id=690926