manager: return ACCESS DENIED for most open-reauth-channel failures
gnome-shell only falls back to opening a new session if the reauth
channel fails to open because of denied access.  For XDMCP we currently
fail for other reasons.

This commit returns ACCESS DENIED for those cases as well, so the
intended fallback happens.

Based on a patch by Jeremy White <jwhite@codeweavers.com>

https://bugzilla.gnome.org/show_bug.cgi?id=690926