slave: Pass session id to unlocking code if we know it
We currently figure out which session to unlock on-the-fly based on
the username that was just reauthenticated.  This doesn't work, for
XDMCP sessions, under logind, since they don't have a seat. There's
no reason to try to deduce the session, though.  We already know it,
so we might as well pass it down.

This fixes unlocking for XDMCP.

https://bugzilla.gnome.org/show_bug.cgi?id=690926