Don't enable debugging by default if the version is "unstable"
I'm open to negotiation for this, but basically...gdm is at
present second only to the kernel in the sheer amount of spew
it emits - when the version is "unstable".  But it always will
be for Continuous.

Nowadays of course, I think Continuous helps act as a testing system
for GDM that obviates a lot of the need for lots of debugging info for
gdm during "unstable" cycles.

I suspect that we will over time add other informational messages at
strategic points by default; this patch is not the end of the story.
It's a new beginning.

https://bugzilla.gnome.org/show_bug.cgi?id=715037