move over to using 'dconf compile' and file-db
Instead of installing keyfiles in /etc and running 'dconf update' to
generate the database in place (or requiring package system triggers to
do so for us), just build the database in-tree using the new 'dconf
compile' command.  We can install that in our pkgdatadir.

dconf now also looks in XDG_DATA_DIRS in addition to /etc when finding
profiles, so install ours there too.

This makes gdm /etc-clean with respect to dconf.

If users want to make additional customisations to the gdm login session
then they should create their own dconf database in /etc/dconf and then
replace the 'gdm' profile with one that references their new db (by way
of /etc/dconf/profile/gdm, which is searched first).

https://bugzilla.gnome.org/show_bug.cgi?id=722241