Stop using LDFLAGS for libraries
Originally, EXTRA_*_LIBS was introduced as a way to get certain
libraries to the front of the line when invoking the linker in order to
control which version of a library we got.  See bug 85785 for the
history there.

This is probably no longer needed, so officially declare that the only
thing that the EXTRA_ variables mean is "not from pkg-config".

Move all of the EXTRA_*_LIBS from LDFLAGS to _LDADD, along with some
other things that were improperly in LDFLAGS.

https://bugzilla.gnome.org/show_bug.cgi?id=724283