launch-environment: don't disable GVFS
As an optimization we currently disable gvfs in greeter sessions
(since they'll never need it).

Now that we launch initial-setup in the same sort of environment,
we're running into trouble because it relies on GVFS for fetching
remote avatars.

This commit eliminates the optimzation. The other option would be to
disable GVFS for greeters, but enable it for initial-setup. In order
to keep things simple, we forgo that for now, but may reconsider at
a later time.

https://bugzilla.gnome.org/show_bug.cgi?id=725584