manager: Rename the "user session" or "session" to the "seed session"
A greeter environment always has two sessions: the greeter session, and
the "user session".

The "user session" tracks the login PAM conversation. PAM conversations
need to be created without any user, so the name "user session" is a bit
of a misnomer.

The "user session" is created when the greeter is created, and exists
when we're sitting at the username prompt. When the username is entered,
we then start a session worker. When the verification conversation is
finished, we then log the user in. The same exact GdmSession follows
into the user session.

Traditionally, the user session is started on the same X server as the
greeter, so the GdmDisplay and GdmSlave all go with it. With the new
display server management, we need to make the GdmSession follow into
the user session created for the login, and then start a *new* user
session for the next user that logs in on the same greeter.

The name "seed session" tries to capture this. The GdmSession starts out
as a little seed, and with a username and password, will one day blossom
into a beautiful little user session which will one day live on its own,
free from the greeter that gave it life and nourished it.

It's the circle of life, and it moves us all.

https://bugzilla.gnome.org/show_bug.cgi?id=726380