launch-environment: disconnect session signal handlers when destroyed
The launch environment's session can actually live longer than the
environment itself.  This is because the session creates internals
"conversation" objects that take a reference on the session, which
sometimes keeps it on life support.

This commit makes sure the session signal handlers for the launch
environment don't get run after the launch environment is destroyed.

https://bugzilla.gnome.org/show_bug.cgi?id=726380