manager: update session-id when starting user session
In order for session unlocking to work properly, the session-id
property of the display object needs to match the session running
on the display.  The display object's session-id property is bound
to the slave's session-id property, so this commit makes sure
the slave's session-id is updated to the user's session's id once
the session is started.

https://bugzilla.gnome.org/show_bug.cgi?id=726380