manager: Don't reuse the same X server for alternate display modes
GNOME is going to need a mode of operation in GDM where the session
manages its own display server. In this mode of operation, we won't
morph the login screen into a user session by reusing the same X server.

Instead, we reset the login screen to prepare it for a future login,
and then let GNOME or the session worker activate its own session.

This commit adds the prerequisite work needed to gdm-manager.c, but
only as dead code, since the way to "turn it on" is stubbed out to
always reuse the existing server.

https://bugzilla.gnome.org/show_bug.cgi?id=726380