manager: wait until autologin is initiated before marking it so
When the slave was merged to the main daemon process, the code for
doing autologin accounting was accidentally moved earlier in the
autologin process.  This caused the greeter to create a login
screen and do autologin at the same time.

This commit defers the accounting until after the operation is
fully initiated.

https://bugzilla.gnome.org/show_bug.cgi?id=726813