ConsoleKit: a pair of tweaks
It looks like there were a couple of breaks when porting away from
ck-connector.

The first issue is that consolekit uses some hard-coded rules to decide
if a given session is associated with the hard-coded 'Seat1' or not.
One of these rules is the "x11-display-device" property being set.  We
were setting the "display-device" property instead, which is not for
graphical sessions.  This was causing a new seat to be dynamically
created each time gdm started which was causing us to send
ActivateSession requests to the wrong seat.

The second issue is that ActivateSession will "fail" with "Session
already active" errors.  Ignore those.

https://bugzilla.gnome.org/show_bug.cgi?id=727072