manager: be more accepting of callers with uids different from their session
If a user runs su in their session, that switched user will now be
running in a session that doesn't belong to it.  GDM won't allow a user
access to the worker process associated with the session in this case.
Instead, it will try to create a temporary just-in-time reauthentication
channel so reauthentication can happen without having the user talking to
another user's worker. Unfortunately, a logic error in the code means,
the user won't access to its own just-in-time channel.

This commit fixes that.

https://bugzilla.gnome.org/show_bug.cgi?id=727982