session: make sure session display mode is set even if session isn't explicitly picked
Right now we we only set the display mode when a user picks a session
from the session list.  That means the wrong display mode will get used
if wayland is saved from a previous log in.

This commit makes sure the display mode gets set as soon as the user
credentials are established, since at the point log in is imminent.

https://bugzilla.gnome.org/show_bug.cgi?id=728204