worker: reuse user's VT if available
When a user logs out, their session goes into the closing
state until all processes from the session exit. The X server
won't exit while the session is in the closing state, but will
instead reset itself.  When the user logs in again, we need to
select the same VT so the already running X server can properly
manage the new session.

This commit checks for the above scenario and tries to account
for it.