gdm-session: always pass down user saved language
There's a bug when changing the language of the login screen to a different
one than the user session's language, which causes the new language to be
used both in the login screen AND the user session if you simply logout and
then re-login without rebooting. Restarting the machine after that point
normalizes things, getting the new language used only in the login screen
but not in the user session.

Making sure that the user language is always saved by the session seems
without checking whether is the same than the default language fixes this.

Based on a patch by Felipe Erias Morandeira <femorandeira@igalia.com>

https://bugzilla.gnome.org/show_bug.cgi?id=786422