launch-environment: explicitly kill worker on stop even if session is running
Right now, if the launch environment is stopped while it's running the
whole process group of the session is killed. The theory is that doing
this will kill both the worker and the dbus-daemon (and anything else).

The problem is, now that gdm-simple-slave is gone, the worker is not part
of the same process group as its children, so killing the process group
of the session leaves the worker alive.

This commit makes sure the worker is always explicitly killed.

https://bugzilla.gnome.org/show_bug.cgi?id=729727