transient-display: don't call unmanage from finish
GdmTransientDisplay has this code:

        /* we don't restart/remanage transient displays */
        gdm_display_unmanage (display);

But the display factory already makes sure transient displays aren't
restarted. Furthermore, calling unmanage from finish changes the status
of the display to UNMANAGED instead of FINISHED which causes an
assertion to get blown when the display is disposed if the user logouts
out when user switching.

This commit removes the code since it's unnecessary and actively
harmful.

https://bugzilla.gnome.org/show_bug.cgi?id=728488