worker: don't block SIGUSR1
When the slaves were folded into the main manager process, we had
to add support for detecting multiple simultaneous X servers starting
up.

We did that by creating a detected thread to listen for SIGUSR1 and
then examining the siginfo to find out which X server is ready.

In order to ensure only the dedicated thread received the SIGUSR1
signals from the X server we blocked SIGUSR1 from the main thread.

That blocked signal is inherited to the worker processes and the
session.  Some programs depend on SIGUSR1 but don't explicitly unblock
SIGUSR1 at start up, since the signal is presumed to be unblocked out
the box.

This commit makes sure SIGUSR1 is unblocked before starting the session
to keep these programs functioning correctly.

https://bugzilla.gnome.org/show_bug.cgi?id=731228