session: disconnect dbus connection "closed" handler when no longer needed
We temporarily maintain a reference to a dbus connection to the worker
process in a linked list, while we wait for the worker to register.  If
we need to close the connection before the worker registers, then we
drop the held reference by way of a "closed" signal handler on the
connection object.

Once the worker registers, we remove the connection from the linked
list, and transfer ownership of the connection to a state variable in
the now running conversation object.  When performing the reference
transfer, we neglect to the disconnect the "closed" handler.

This commit makes sure the handler gets disconnected appropriately, so
it doesn't get called inappropriately.

https://bugzilla.gnome.org/show_bug.cgi?id=727589