session: make sure worker manager interface is freed up
Right now we unref the worker manager interface in one function:

stop_conversation_now

We neglect to unref it in stop_conversation and, of course, it's
owned by the conversation object so it should get freed when the
conversation is freed.

This commit changes the stop functions to deactivate the interface
in stop, changes the free function to drop the reference to the interface.

https://bugzilla.gnome.org/show_bug.cgi?id=727589