daemon: allow multiple time logins to work
GDM currently lumps timed login with autologin when deciding whether or
not to skip it, if it's already happened once this boot.

That's the wrong behavior, and indeed will create a confused experience,
because it's not what the login screen provided by gnome-shell
expects.

This commit allows timed login to work after the user logs out.

https://bugzilla.gnome.org/show_bug.cgi?id=732774