session: run logind managed displays through session wrappers
Now that we have a utility for starting a wayland session
(and registering it with GDM) we should use the tool.

Now that we have a utility for starting an X server from within the
user session, we can support the LOGIND_MANAGED display mode for X sessions,
in a similar way to how we support wayland sessions on a new vt.

This commit adds the plumbing necessary to use the new tools.