manager: handle autologin directly
Right now we call a "set_up_greeter_session" method on the display
object to to find out the autologin username.  This is weird because
there's no greeter session when performing an autologin.  Also, the
display delegates to the slave, and we want to get rid of the slave.
Furthermore, the manager handles most of the autologin implementation
already, it's just the "get the username" part (which is always "root"),
that it delegates.

This commit changes the manager to just handle the autologin process
completely by itself.