local-display-factory: drop some superfluous code
The store_display function takes a display number as
an argument, it then uses that display number to reserve
a slot in the "used display numbers" hash table.

There's no reason for it to do this though, because the
slot is already reserved when the next available display
number is looked up in take_next_display_number.

This commit drops the superfluous code.

https://bugzilla.gnome.org/show_bug.cgi?id=744764