manager: rename seed session to embryonic user session
Previously the seed session was called the "user session".
This was confusing because it exists before the user logs in,
so it got renamed to "seed session".  This is confusing because
it's not immediately clear what a seed session is.

It's a lot of typing, but I'm going to go with "embryonic user session"
now. Then it's clear it's destined to be a user session.