manager: don't pass display to functions that fetch timed/autologin details
The only reason we need the display for autologin is to know if it's
running on seat0 or not.  Going forward we won't necessarily have a
display up front, so drop the parameter, and just check the seat
in the caller.

For symmetry, drop the display paramater timed login, too, and do
the display related checks in the lone caller.

https://bugzilla.gnome.org/show_bug.cgi?id=746492