manager: set display name on session object at registration time
When the X server used at login time registers with GDM, GDM tries
to write a wtmp session record for it.

Now that the X server is started in the session, we don't know
the display name of the X server up front and so don't have the
display name attached to the session object. The wtmp record writing
code relies on getting the display name from the session object, and
so it fails.

We do know the display name at registration time, from the details
passed to the registration function.

This commit makes sure to attach the display name to the session object
as soon as the display is registered before writing the wtmp record.

https://bugzilla.gnome.org/show_bug.cgi?id=747169