manager: find session at registration time
We try to look up the session at registration time to add wtmp
records for it.  We fail to actually find the session, though,
because we're using the "embryonic-user-session" object data
key, which is only non-NULL when the user session is still
getting setup.

This commit changes the registration code, to instead, fetch
the session straight from the manager object.

https://bugzilla.gnome.org/show_bug.cgi?id=747169