legacy-display: ensure X11 display device is propagated to launch environment
Once the server associated with the login screen session is ready, we
query its display device for ConsoleKit.  This device needs to get
propagated to the session to ensure ConsoleKit can track session
activeness properly.

This commit makes sure the display device is plumbed from the GdmServer
object to the GdmLaunchEnvironment object where it gets used by the
the login session (and subsequently the user session).

https://bugzilla.gnome.org/show_bug.cgi?id=747351