manager: fix monitor hotplug segfault
commit e5a0e92f59e256edc6489f2234fbe54c25ba9743 introduced a way to
find a user session associated with a display object.

That function has a bug in it, where it skips every even
registered user session because it follows the next pointer
twice per iteration of the loop.

This can cause a crash on monitor hotplug, and in other scenarios
if there are an odd number of user sessions (since the
terminating NULL will be even and skipped over).

https://bugzilla.gnome.org/show_bug.cgi?id=749987