manager: filter out wayland user sessions if X11 login screen
We use wayland by default on the login screen, so if we're running
in an X11 session, then the presumption is that we're incapable of
using wayland. If we're incapable of using wayland sessions then
we should ignore wayland sessions read from user configuration and
fall back to non-wayland defaults.