libgdm: filter out wayland sessions from X11 login screen
We use wayland by default on the login screen, so if we're running
in an X11 session, then the presumption is that we're incapable of
using wayland. If we're incapable of using wayland sessions then
we shouldn't present those sessions to the user in the session list
either.

This commit makes sure to avoid showing wayland based sessions
in the session list if the login screen is running on X11.

https://bugzilla.gnome.org/show_bug.cgi?id=757715