local-display-factory: call sync function after display finishes
Right now if a local display finishes we create a new one of the
same type to replace it.  But if a user display finishes we don't
want to create a new user display, we want to create greeter
display.  Furthermore if the user session is X11, we still want the
login screen that comes up to be wayland by default.

This commit changes the code to just call sync_seats after a display
finishes, since sync_seats is where the logic for where the login
screen display information is stored.

sync_seats is idempotent anyway, and will only create displays for
seats that need them, so it's safe to call it any time any local
display finishes.

https://bugzilla.gnome.org/show_bug.cgi?id=749418