manager: only claen up stub greeter display if it's actually a stub
commit 2774c7e43b9fdf5e5e59ef1f53ae7ba29f4aa23c fixed a leak in the
case we're doing autologin, but it also unconditionally cleans up
the display in the non-autologin case, too.

At the moment, we don't want to clean up the greeter in that case,
since it causes unnecessary flicker and slows down fast user switching.
( though this behavior may get changed again pending the outcome of
  bug 747339 )

This commit makes sure we only clean up the leaker greeter display, if
there is a leaked greeter display (namely, the autologin case)

https://bugzilla.gnome.org/show_bug.cgi?id=749418