manager: don't use wayland for autologin if wayland disabled
These days we default to using wayland user sessions if the
unless greeter isn't wayland.  But in the case of automatic
login there is no greeter.

This commit makes automatic login ignore wayland sessions if
wayland is disabled in the config file. This gives us symetry
with the greeter case, and provides a way for users to opt
out of wayland.

https://bugzilla.gnome.org/show_bug.cgi?id=757715