session: make sure greeter is GNOME session even if gnome isn't installed
We figure out the XDG_CURRENT_DESKTOP by looking at the session that's
use for login.  For the greeter, we don't use a session, at all, but
instead a directly specified program.  Still, we're determining the
XDG_CURRENT_DESKTOP for the greeter based on what default session
gets picked.  On KDE only setups, the greeter will get marked as
a KDE session which will prevent gnome-settings-daemon from starting
properly.

This commit changes the code so that greeter sessions, don't rely
on the default session desktop file.

https://bugzilla.gnome.org/show_bug.cgi?id=760943