gdm-session: set PAM_TTY when initialising pam
Temporarily set this to the VT that the login screen is currently active on so that pam modules that require this
information early can have access to it. Debian/Ubuntu use this to detect if the login is coming from a local tty.
PAM_TTY will get reset to the users VT right before the user session is opened.

https://bugzilla.gnome.org/show_bug.cgi?id=764669