local-display-factory: disconnect signal handlers when factory is disposed
There's the potential for a crash in the shutdown path after the
factory is disposed, since we fail to disconnect signal handlers to
the displays / display store at factory dispose time.

This commit changes the g_signal_connect to g_signal_connect_object, to
avoid any potential for crash.

(this is a fix noticed when reading through the source.  It's tangentially
related to a discussion on irc for a different bug)