xdmcp-display-factory: notify remote display when session ended
gnome-shell and the session dbus daemon don't automatically exit
when gnome-session does.  They, instead, wait for the display to
exit or regenerate. If the display is remote, that won't happen
until the keep alive timeout.

This commit changes GDM to explicitly notify the remote display
when the session is over, so that it can regenerate immediately.

https://bugzilla.gnome.org/show_bug.cgi?id=776059