manager: make sure we retain ignore-wayland on second login
The intention of the code is to only allow wayland login for
user sessions if the greeter session is also wayland.

Right now, that intention is only honored the first time a user
logs in.

This commit corrects the problem, to make sure sure we always avoid
wayland if the greeter session avoided wayland.

https://bugzilla.gnome.org/show_bug.cgi?id=775659