session: update session type when session is read
We currently update the session type (to either wayland or
x11) when we first start the PAM conversation and later when
the username is set (if the user is not set at the same time as
the PAM conversation).

There's a race that means the session won't necessarly have been
read from accountsservice at these points.

This commit changes the code to instead update the session type
in result to the session actually getting read.

https://bugzilla.gnome.org/show_bug.cgi?id=781825