session: ensure wayland paths get searched even with --disable-user-display-server
get_system_session_dirs is function that says which directories to
search for session desktop files.  The returned directories may
be xsession dirs, wayland-session dirs, or both, depending on
configuration.

A bug in this function means an unused NULL sentinal value from the
xsession list of directories is getting copied into the middle of
the output list, when --disable-user-display-server is passed to
configure.

This commit corrects the bug, by dropping the unused sentinal value.

Launchpad-Bug: 1704050

https://bugzilla.gnome.org/show_bug.cgi?id=784891