daemon: add ChoiceList PAM extension
This commit adds one PAM extension, a "Choice List" using the
new PAM_BINARY_PROMPT protocol added in the previous commit.  The
PAM module sends a list of (key, row text) pairs, and GDM ferries
the request to gnome-shell using a new user verifier sub-interface.

gnome-shell should present the list to the user and pass back the
corresponding key, which GDM ferries back to the PAM module.

Note this commit is only the daemon side. A subsequent commit will
add the libgdm API needed for gnome-shell to actually deal with
this new PAM extension.

https://bugzilla.gnome.org/show_bug.cgi?id=788851