gdm-{wayland,x}-session: only omit unset session variables from environment
The display server launcher wrapper scripts try to unset any stale
environment variables imported from systemd.

There is a possibility, some of these stale variables, however, could be
"refreshed" by gdm in the current session.  If that's the case, we
shouldn't clear them.

This reworks the unset calls into a loop, and guards each unset call
with a check to make sure it's not in the current environment.

https://bugzilla.gnome.org/show_bug.cgi?id=785918