manager: rename embryonic-user-session to user session
The user session exists and gets used after it's fully grown,
(when reauthenticating), so calling it an embryonic session
is weird.  The word embryonic is weird anyway.  Previously,
we called it a seed session, and that was weird, too, for the
same reason.  Before that we just called it a user session,
so let's come around full circle and name it user session again.

https://bugzilla.gnome.org/show_bug.cgi?id=784555