manager: make sure a session's display knows about the session
When logging into a session that runs on another VT, we create a
new display object for it.  That display object needs to know
about the session, so we can find the session when reauthenticating
against the display.

The code that links the two together was missing.  This commit
adds it.

https://bugzilla.gnome.org/show_bug.cgi?id=784555