manager: don't allow autologin from transient displays
In theory, we're only only supposed to allow autologin
the first time a session is run, but we only count a
session run, once it's finished.  This means that if a
user creates a transient session to user switch, before
they've logged out the first time at boot up, that
transient session will begin autologin as well (which
actually gets treated as an auto unlock).

This commit makes sure autologin is only ever run on
the initial display.

https://bugzilla.gnome.org/show_bug.cgi?id=783779