gdm-session: Calculate fallback session name without ".desktop"
When the user has not ever selected a session from the selector, we will
log them into the `fallback' session, which is calculated by looking at
all installed sessions and picking the first one in the list.

There is a bug that the visually selected session presented in the
greeter can be different from the fallback session calculated inside
GDM. This happens because the two sort sessions differently.

Inside GDM we calculate the fallback session by comparing the full
basename of the file:

>>> GLib.strcmp0("gnome-xorg.desktop", "gnome.desktop")
-1

Whereas gnome-shell calls gdm_get_session_ids(), which returns IDs
without the ".desktop" extension. Those are then sorted:

>>> GLib.strcmp0("gnome-xorg", "gnome")
45

which gives a different ordering.

Fix this bug by calculating the fallback session in the same way, by
removing ".desktop" from the filename.

https://bugs.launchpad.net/ubuntu/+source/gnome-session/+bug/1705157

https://bugzilla.gnome.org/show_bug.cgi?id=787304