xdmcp-display-factory: more signal prototype fixing
commit 2a3d5a7c402534c3611956fed930394a51cb00c5 attempted to fix
the signal prototype of on_client_disconnected to match the marshalled
arguments.  Unfortunately, the handler is used for two different
signals:

1) when a user chooses "Disconnect" from the chooser menu
2) when a user closes the Xephyr window (or turns off the thin client
or whatever)

The signals have slightly different prototypes, so commit 2a3d5a7c
fixed 2), but broke 1).

This commit massages the signal connections and handler signature to
work for both cases.

https://bugzilla.gnome.org/show_bug.cgi?id=787987