daemon/gdm-manager.c: quit plymouth when xdmcp is the only allowed connection.
gdm is responsible to kill plymouth by spawning the "plymouth quit"
subprocesses in gdm-manager.c. The current code pathes of quiting
plymouth can never be reached when xdmcp is the only connection
allowed. Consequently in the case of

    !show_local_greeter && xdmcp_enabled

the plymouth-quit-wait.service will never quit and the login prompt
will not popup without manual interference. This issue could be
more obviously observed when a downstream like openSUSE which
allows a customized sysconfig to switch the corresponding two
options on a headless server (s390), where the setup is usually:

    DISPLAYMANAGER_REMOTE_ACCESS="yes"
    DISPLAYMANAGER_STARTS_XSERVER="no"

The proposed patch handles this edge case by quit plymouth immediately
when the condition is detected.

https://bugzilla.gnome.org/show_bug.cgi?id=795120