libgdm: Don't unref a connection that's in use
If an async task tries to reuse an open connection,
it erroneously explicitly unrefs it. That is incorrect,
because there are weak references in use to handle
disposing the connection when its no longer in use.

This commit makes sure the local connection object
in open_connection is nullified so the connection
doesn't get autofree'd.

https://bugzilla.gnome.org/show_bug.cgi?id=795940
(cherry picked from commit 4654042d22c71da599d60a7e2f0570acf3d4a166)