libgdm: don't keep manager proxy around longer than we need it
Right now we keep the manager proxy alive long after we need it.
It doesn't get cleared until one of the other proxies go away.

That is not only unnecessary but illogical and confusing.

This commit changes the manager proxy to be transient—only alive
long enough to get what we need from it.

https://bugzilla.gnome.org/show_bug.cgi?id=795940